+++ packages/agent/src/actor.ts
@@ -658,7 +658,7 @@ export type ManagementCanisterRecord = _SERVICE;
 export function getManagementCanister(config: CallConfig): ActorSubclass<ManagementCanisterRecord> {
   function transform(
     _methodName: string,
-    args: Record<string, unknown> & { canister_id: string }[],
+    args: Record<string, unknown> & { canister_id: string; target_canister?: unknown }[],
   ) {
     if (config.effectiveCanisterId) {
       return { effectiveCanisterId: Principal.from(config.effectiveCanisterId) };
@@ -667,6 +667,9 @@ export function getManagementCanister(config: CallConfig): ActorSubclass<Managem
     let effectiveCanisterId = Principal.fromHex('');
     if (first && typeof first === 'object' && first.canister_id) {
       effectiveCanisterId = Principal.from(first.canister_id as unknown);
+    }
+    if (first && typeof first === 'object' && first.target_canister) {
+      effectiveCanisterId = Principal.from(first.target_canister);
     }
     return { effectiveCanisterId };
   }